Perform some code refactoring
