+++ kernel/src/arch/x86_64/mod.rs
@@ -414,7 +414,7 @@ fn _start(argc: isize, _argv: *const *const u8) -> isize {
         .get()
         .map_or(false, |c| c.mode == crate::cmdline::Mode::Client)
     {
-        let _ = spin::lazy::Lazy::force(&rackscale::RPC_CLIENT);
+        let _ = spin::lazy::Lazy::force(&rackscale::client::RPC_CLIENT);
     }
 
     // Bring up the rest of the system (needs topology, APIC, and global memory)
