+++ kernel/src/arch/x86_64/rackscale/controller.rs
@@ -1,20 +1,31 @@
 // Copyright © 2021 University of Colorado. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR MIT
+
 use alloc::boxed::Box;
 use alloc::sync::Arc;
 use alloc::vec::Vec;
 use core::cell::Cell;
 use fallible_collections::FallibleVecGlobal;
 use hashbrown::HashMap;
+use lazy_static::lazy_static;
+use log::{debug, error, warn};
 use smoltcp::time::Instant;
+use static_assertions as sa;
 
 use rpc::api::RPCServer;
-use rpc::rpc::RPCType;
+use rpc::rpc::{ClientId, RPCType};
 use rpc::server::Server;
 
 use crate::arch::debug::shutdown;
 use crate::arch::rackscale::dcm::*;
+use crate::cmdline::Transport;
+use crate::error::KError;
+use crate::fs::{cnrfs, NrLock};
 use crate::memory::backends::AllocatorStatistics;
+use crate::memory::mcache::MCache;
+use crate::memory::LARGE_PAGE_SIZE;
+use crate::nr;
+use crate::process::Pid;
 use crate::transport::ethernet::ETHERNET_IFACE;
 use crate::transport::shmem::create_shmem_manager;
 use crate::ExitReason;
@@ -22,6 +33,21 @@ use crate::ExitReason;
 use super::*;
 
 const PORT: u16 = 6970;
+
+/// A cache of pages
+/// TODO: think about how we should constrain this?
+///
+/// Used to allocate remote memory (in large chunks)
+pub(crate) type FrameCacheMemslice = MCache<2048, 2048>;
+sa::const_assert!(core::mem::size_of::<FrameCacheMemslice>() <= LARGE_PAGE_SIZE);
+sa::const_assert!(core::mem::align_of::<FrameCacheMemslice>() <= LARGE_PAGE_SIZE);
+
+// Mapping between local PIDs and remote (client) PIDs.
+// Using (ClientId, Pid) works as long as each ClientId has it's own
+// unique Pid space.
+lazy_static! {
+    static ref PID_MAP: NrLock<HashMap<(ClientId, Pid), Pid>> = NrLock::default();
+}
 
 /// Test TCP RPC-based controller
 pub(crate) fn run() {
@@ -146,4 +172,56 @@ fn register_rpcs(server: &mut Box<dyn RPCServer>) {
     server
         .register(KernelRpc::RequestCore as RPCType, &CORE_HANDLER)
         .unwrap();
+}
+
+// Lookup the local pid corresponding to a remote pid
+pub(crate) fn get_local_pid(client_id: ClientId, remote_pid: usize) -> Result<usize, KError> {
+    {
+        let process_lookup = PID_MAP.read();
+        let local_pid = process_lookup.get(&(client_id, remote_pid));
+        if let Some(pid) = local_pid {
+            return Ok(*(local_pid.unwrap()));
+        }
+    }
+
+    // TODO: will eventually want to delete this logic, as we should create
+    // mapping on process creation.
+    warn!(
+        "Failed to lookup remote pid {}:{}, will register locally instead",
+        client_id, remote_pid
+    );
+    register_pid(client_id, remote_pid)
+}
+
+// Register a remote pid by creating a local pid and creating a remote-local PID mapping
+pub(crate) fn register_pid(client_id: ClientId, remote_pid: usize) -> Result<usize, KError> {
+    crate::nr::NR_REPLICA
+        .get()
+        .map_or(Err(KError::ReplicaNotSet), |(replica, token)| {
+            let response = replica.execute_mut(nr::Op::AllocatePid, *token)?;
+            if let nr::NodeResult::PidAllocated(local_pid) = response {
+                // TODO: some way to unwind if fails??
+                match cnrfs::MlnrKernelNode::add_process(local_pid) {
+                    Ok(_) => {
+                        // TODO: register pid
+                        debug!("register_pid about to get PID_MAP");
+                        let mut pmap = PID_MAP.write();
+                        pmap.try_reserve(1)?;
+                        debug!(
+                            "Mapped remote pid {} to local pid {}",
+                            remote_pid, local_pid
+                        );
+                        pmap.try_insert((client_id, remote_pid), local_pid)
+                            .map_err(|_e| KError::FileDescForPidAlreadyAdded)?;
+                        Ok(local_pid)
+                    }
+                    Err(err) => {
+                        error!("Unable to register pid {:?} {:?}", remote_pid, err);
+                        Err(KError::NoProcessFoundForPid)
+                    }
+                }
+            } else {
+                Err(KError::NoProcessFoundForPid)
+            }
+        })
 }
