+++ kernel/src/arch/x86_64/rackscale/fileops/getinfo.rs
@@ -11,9 +11,9 @@ use rpc::RPCClient;
 use crate::fallible_string::TryString;
 use crate::fs::cnrfs;
 
-use super::super::get_local_pid;
 use super::super::kernelrpc::*;
 use super::FileIO;
+use crate::arch::rackscale::controller::get_local_pid;
 
 pub(crate) fn rpc_getinfo<P: AsRef<[u8]> + Debug>(
     rpc_client: &mut dyn RPCClient,
