+++ kernel/src/arch/x86_64/rackscale/fileops/mkdir.rs
@@ -15,7 +15,7 @@ use rpc::RPCClient;
 
 use super::super::kernelrpc::*;
 use super::FileIO;
-use crate::arch::rackscale::get_local_pid;
+use crate::arch::rackscale::controller::get_local_pid;
 use crate::fallible_string::TryString;
 use crate::fs::cnrfs;
