+++ kernel/src/arch/x86_64/rackscale/fileops/rename.rs
@@ -14,7 +14,7 @@ use rpc::RPCClient;
 
 use super::super::kernelrpc::*;
 use super::FileIO;
-use crate::arch::rackscale::get_local_pid;
+use crate::arch::rackscale::controller::get_local_pid;
 use crate::fallible_string::TryString;
 use crate::fs::cnrfs;
