+++ kernel/src/arch/x86_64/rackscale/fileops/rw.rs
@@ -14,7 +14,7 @@ use crate::fs::fd::FileDescriptor;
 
 use super::super::kernelrpc::*;
 use super::FileIO;
-use crate::arch::rackscale::get_local_pid;
+use crate::arch::rackscale::controller::get_local_pid;
 
 #[derive(Debug)]
 pub(crate) struct RWReq {
